--- src/factories/pair/uniswap-pair-contract.factory.public.ts
@@ -1,9 +1,0 @@
-import { ChainId } from '../../enums/chain-id';
-import { EthersProvider } from '../../ethers-provider';
-import { UniswapPairContractFactory } from './uniswap-pair-contract.factory';
-
-export class UniswapPairContractFactoryPublic extends UniswapPairContractFactory {
-  constructor(chainId: ChainId, providerUrl?: string | undefined) {
-    super(new EthersProvider(chainId, providerUrl));
-  }
-}
