+++ src/index.ts
@@ -15,7 +15,7 @@ export {
 } from './factories/pair/models/uniswap-pair-contexts';
 export { UniswapPairSettings } from './factories/pair/models/uniswap-pair-settings';
 export { UniswapPair } from './factories/pair/uniswap-pair';
-export { UniswapPairContractFactoryPublic } from './factories/pair/uniswap-pair-contract.factory.public';
+// export { UniswapPairContractFactoryPublicV2 } from './factories/pair/uniswap-pair-contract.factory.public.v2';
 export { UniswapPairFactory } from './factories/pair/uniswap-pair.factory';
 export { RouteQuote } from './factories/router/models/route-quote';
 export { UniswapRouterContractFactoryPublic } from './factories/router/v2/uniswap-router-contract.factory.public.v2';
